Receiver notifies sender with reasons for discarding envelopes (#23)
* Receiver notifies sender with reasons for discarding envelopes

On a sender side we want to know if some messages were discarded. For this
purpose new message type was added - messagesResponse. This message type is versioned,
in order to make it easier transition to new message response, if we will want to extend it in near
future.

The response itself is formed from hash of envelopes received in a single batch, and errors
during processing of envelopes. Each error has a code and description, the reason for adding a code
is to provide some kind of user-friendly description to a client explaining why certain envelopes
were not yet delivered. Description will be simply logged on a sender side.

The followup for this change will be an extention on status-go side that will send particular envelope
errors to a client (e.g. status-react).

* Store errors as array and cover new changes with tests